refactor(expressions): Remove v1 SPEL code
- Making v2 as the default engine for expressions
- More improvements on the way
